move JSON serialization into json.rs module; remove JSON serialization for ErgoBoxCandidate;
