+++ bindings/ergo-wallet-wasm/tests/test_transaction.js
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import {
   Address, Wallet, UnspentBoxes, TxDataInputs, ErgoBoxCandidate, Contract,
-  TxOutputs, ErgoStateContext
+  TxOutputCandidates, ErgoStateContext
 } from '../pkg/ergo_wallet_wasm';
 
 it('new signed transaction', async () => {
@@ -13,7 +13,7 @@ it('new signed transaction', async () => {
   const send_change_to = Address.from_testnet_str('');
 
   let outbox = new ErgoBoxCandidate(1, 0, Contract.pay_to_address(recipient));
-  let tx_outputs = TxOutputs.from_boxes([outbox]);
+  let tx_outputs = new TxOutputCandidates(outbox);
   let dummy_ctx = ErgoStateContext.dummy();
   let wallet = Wallet.from_mnemonic("", "");
   expect(() => wallet.new_signed_transaction(dummy_ctx,
